firstbutton/src/backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
google-generativeai
google-auth-oauthlib
google-api-python-client
python-dotenv
Pillow
python-multipart
boto3
